Rename mpmpid to pid
I only had it named mpmpid to make it clear it was not mine. However, I still have a package with the same name in my repo, so it would be nice if the official one was named pid. I will update my programs to use the official one soon.